rename the test directories
