+++ scripts/concent_integration_tests/nodes/provider/debug.py
@@ -1,17 +1,17 @@
 #!/usr/bin/env python
 """
 
-Regular, unmodified Requestor node, running with DEBUG loglevel
+Regular, unmodified Provider node, running with DEBUG loglevel
 
 """
 
 import sys
-from scripts.concent_node_tests import params
+from scripts.concent_integration_tests import params
 
 sys.path.insert(0, 'golem')
 
 from golemapp import start  # noqa: E402 module level import not at top of file
 
-sys.argv.extend(params.REQUESTOR_ARGS_DEBUG)
+sys.argv.extend(params.PROVIDER_ARGS_DEBUG)
 
 start()
