+++ scripts/concent_integration_tests/nodes/provider/regular.py
@@ -1,17 +1,17 @@
 #!/usr/bin/env python
 """
 
-Regular, unmodified Requestor node
+Regular, unmodified Provider node
 
 """
 
 import sys
-from scripts.concent_node_tests import params
+from scripts.concent_integration_tests import params
 
 sys.path.insert(0, 'golem')
 
 from golemapp import start  # noqa: E402 module level import not at top of file
 
-sys.argv.extend(params.REQUESTOR_ARGS)
+sys.argv.extend(params.PROVIDER_ARGS)
 
 start()
