+++ scripts/concent_integration_tests/nodes/requestor/debug.py
@@ -1,17 +1,17 @@
 #!/usr/bin/env python
 """
 
-Regular, unmodified Provider node
+Regular, unmodified Requestor node, running with DEBUG loglevel
 
 """
 
 import sys
-from scripts.concent_node_tests import params
+from scripts.concent_integration_tests import params
 
 sys.path.insert(0, 'golem')
 
 from golemapp import start  # noqa: E402 module level import not at top of file
 
-sys.argv.extend(params.PROVIDER_ARGS)
+sys.argv.extend(params.REQUESTOR_ARGS_DEBUG)
 
 start()
