+++ scripts/concent_integration_tests/nodes/requestor/fail_results.py
@@ -9,7 +9,7 @@ import mock
 import sys
 import time
 
-from scripts.concent_node_tests import params
+from scripts.concent_integration_tests import params
 
 sys.path.insert(0, 'golem')
