+++ scripts/concent_integration_tests/nodes/requestor/no_ack_rct.py
@@ -11,7 +11,7 @@ import mock
 import sys
 
 from golem_messages.message import RandVal
-from scripts.concent_node_tests import params
+from scripts.concent_integration_tests import params
 
 sys.path.insert(0, 'golem')
