+++ scripts/concent_integration_tests/nodes/requestor/reject_results.py
@@ -9,7 +9,7 @@ iow, always sends `SubtaskResultsRejected`.
 import mock
 import sys
 
-from scripts.concent_node_tests import params
+from scripts.concent_integration_tests import params
 
 sys.path.insert(0, 'golem')
