+++ scripts/concent_integration_tests/tests/test_force_report.py
@@ -1,28 +1,30 @@
 #!/usr/bin/env python
 import sys
 
-from scripts.concent_node_tests import helpers
-from scripts.concent_node_tests.tests.base import NodeTestPlaybook
+from scripts.concent_integration_tests import helpers
+from scripts.concent_integration_tests.tests.base import NodeTestPlaybook
 
 
-class ForceAccept(NodeTestPlaybook):
-    provider_node_script = 'provider/debug'
-    requestor_node_script = 'requestor/no_sra'
+class ForceReport(NodeTestPlaybook):
+    provider_node_script = 'provider/impatient_frct'
+    requestor_node_script = 'requestor/no_ack_rct'
 
     def step_clear_provider_output(self):
         helpers.clear_output(self.provider_output_queue)
         self.next()
 
-    def step_wait(self):
+    def step_wait_task_finished(self):
         concent_fail = helpers.search_output(
             self.provider_output_queue,
-            '.*Concent request failed.*',
+            '.*Concent request failed.*|.*Problem interpreting.*',
         )
 
         if concent_fail:
             print("Provider: ", concent_fail.group(0))
             self.fail()
             return
+
+        super().step_wait_task_finished()
 
     steps = (
         NodeTestPlaybook.step_get_provider_key,
@@ -37,11 +39,11 @@ class ForceAccept(NodeTestPlaybook):
         NodeTestPlaybook.step_create_task,
         NodeTestPlaybook.step_get_task_id,
         NodeTestPlaybook.step_get_task_status,
-        step_wait,
+        step_wait_task_finished,
     )
 
 
-playbook = ForceAccept.start()
+playbook = ForceReport.start()
 if playbook.exit_code:
     print("exit code", playbook.exit_code)
 sys.exit(playbook.exit_code)
